Prepare site for deployment.
