+++ package.json
@@ -26,6 +26,7 @@
     "react-transition-group": "^4.4.5"
   },
   "devDependencies": {
+    "@flydotio/dockerfile": "^0.5.4",
     "@testing-library/jest-dom": "^6.4.2",
     "@testing-library/react": "^14.2.1",
     "@types/better-sqlite3": "^7.6.9",
