+++ src/app/[lang]/account/page.tsx
@@ -9,7 +9,7 @@ import { getActiveSession } from "@/lib/oauth"
 import { Locale, getDictionary } from "@/localization"
 import { cookies } from "next/headers"
 import { redirect } from "next/navigation"
-import { FutureEventItem } from "../events/page"
+import { FutureEventItem } from "../events/future-event-item"
 import { getMaximumSemesterDate, getMinimumSemesterDate, getSemester } from "@/lib/utils"
 import { InputSection } from "@/components/input/input-section"
 import { PortableAccountPage } from "./portable-account-page"
