+++ src/app/[lang]/events/[id]/page.tsx
@@ -5,7 +5,7 @@ import { getActiveSession } from "@/lib/oauth"
 import { Locale, getDictionary } from "@/localization"
 import { cookies } from "next/headers"
 import { redirect } from "next/navigation"
-import { createEventMinAccessLevel, showQRCodeMinAccessLevel } from "../page"
+import { createEventMinAccessLevel, showQRCodeMinAccessLevel } from "@/lib/utils"
 import { isEventInProgress } from "@/lib/utils"
 import { BaseButton } from "@/components/material/base-button"
 import { FilledButton } from "@/components/material/filled-button"
